+++ Source/ACE.Entity/AceObject.cs
@@ -221,6 +221,36 @@ namespace ACE.Entity
             set { SetIntProperty(PropertyInt.Level, value); }
         }
 
+        public uint GeneratorTimeType
+        {
+            get { return GetIntProperty(PropertyInt.GeneratorTimeType) ?? 0; }
+            set { SetIntProperty(PropertyInt.GeneratorTimeType, value); }
+        }
+
+        public uint GeneratorProbability
+        {
+            get { return GetIntProperty(PropertyInt.GeneratorProbability) ?? 0; }
+            set { SetIntProperty(PropertyInt.GeneratorProbability, value); }
+        }
+
+        public uint MaxGeneratedObjects
+        {
+            get { return GetIntProperty(PropertyInt.MaxGeneratedObjects) ?? 0; }
+            set { SetIntProperty(PropertyInt.MaxGeneratedObjects, value); }
+        }
+
+        public uint GeneratorType
+        {
+            get { return GetIntProperty(PropertyInt.GeneratorType) ?? 0; }
+            set { SetIntProperty(PropertyInt.GeneratorType, value); }
+        }
+
+        public uint ActivationCreateClass
+        {
+            get { return GetIntProperty(PropertyInt.ActivationCreateClass) ?? 0; }
+            set { SetIntProperty(PropertyInt.ActivationCreateClass, value); }
+        }
+
         public uint CombatTableId
         {
             get { return GetDataIdProperty(PropertyDataId.CombatTable) ?? 0; }