+++ changelog.md
@@ -1,10 +1,24 @@
 # ACEmulator Change Log
 
+### 2017-06-25
+[OptimShi]
+* Re-enabled the "equiptest" debug command that was inadvertently disabled.
+
 ### 2017-06-24
+[Lidefeath]
+* Remove old creature spawning and generator stuff
+* Implement creature spawning from generator objects, nested or hierarchical generators are still to do
+
 [Ripley]
 * Rebased SQL scripts.
 
 ### 2017-06-23
+[Jyrus]
+* Remove unneeded debug command CreateLifestone
+* Modified PortalObjectFactory class to create specialized portals that are temporary spawns, such as the Humming Crystal Portal
+* Created enums for the weenieclassID's for the special portals
+* Added the missing ObjectType to the CollidableObject class
+
 [ddevec]
 * Fix to teleport freezing characters in place (Wasn't clearing requested location)
 * Fix to occasional crash when moving long distances (Wasn't properly handling landblock transferring)